[P6-141] refactor: my-reservations 페이지 content와 파일 page 분리
